+++ src/DotNext/Runtime/Caching/ConcurrentCache.Eviction.cs
@@ -1,5 +1,4 @@
 using System.Diagnostics;
-using System.Runtime.CompilerServices;
 
 namespace DotNext.Runtime.Caching;
 
@@ -7,36 +6,37 @@ public partial class ConcurrentCache<TKey, TValue>
 {
     private readonly object evictionLock = new();
     private readonly Action<TKey, TValue>? evictionHandler;
-    private readonly Func<KeyValuePair, KeyValuePair?> addCommand, removeCommand, readCommand;
     private int evictionListSize;
     private KeyValuePair? firstPair, lastPair;
 
-    [MethodImpl(MethodImplOptions.NoInlining)]
     private KeyValuePair? OnAdd(KeyValuePair target)
     {
         Debug.Assert(Monitor.IsEntered(evictionLock));
+        Debug.Assert(target.Removed is false);
 
         AddFirst(target);
         evictionListSize += 1;
         return evictionListSize > buckets.Length ? Evict() : null;
     }
 
-    [MethodImpl(MethodImplOptions.NoInlining)]
     private KeyValuePair? OnRemove(KeyValuePair target)
     {
         Debug.Assert(Monitor.IsEntered(evictionLock));
+        Debug.Assert(target.Removed is false);
 
         Detach(target);
         evictionListSize--;
+
         return null;
     }
 
-    [MethodImpl(MethodImplOptions.NoInlining)]
     private KeyValuePair? OnReadLFU(KeyValuePair target)
     {
         Debug.Assert(Monitor.IsEntered(evictionLock));
+        Debug.Assert(target.Removed is false);
 
         var parent = target.Links.Previous?.Links.Previous;
+        Debug.Assert(ReferenceEquals(parent, target) is false);
         Detach(target);
 
         if (parent is null)
@@ -47,10 +47,10 @@ public partial class ConcurrentCache<TKey, TValue>
         return null;
     }
 
-    [MethodImpl(MethodImplOptions.NoInlining)]
     private KeyValuePair? OnReadLRU(KeyValuePair target)
     {
         Debug.Assert(Monitor.IsEntered(evictionLock));
+        Debug.Assert(target.Removed is false);
 
         Detach(target);
         AddFirst(target);
@@ -59,12 +59,18 @@ public partial class ConcurrentCache<TKey, TValue>
 
     private static void Append(KeyValuePair parent, KeyValuePair child)
     {
+        Debug.Assert(ReferenceEquals(parent, child) is false);
+
         child.Links.Previous = parent;
 
-        if ((child.Links.Next = parent.Links.Next) is KeyValuePair childNext)
+        if ((child.Links.Next = parent.Links.Next) is { } childNext)
             childNext.Links.Previous = child;
 
         parent.Links.Next = child;
+
+        Debug.Assert(ReferenceEquals(child, child.Links.Previous) is false);
+        Debug.Assert(ReferenceEquals(child, child.Links.Next) is false);
+        Debug.Assert(ReferenceEquals(child.Links.Next, child.Links.Previous) is false);
     }
 
     private void AddFirst(KeyValuePair pair)
@@ -75,10 +81,18 @@ public partial class ConcurrentCache<TKey, TValue>
         }
         else
         {
+            Debug.Assert(ReferenceEquals(pair, firstPair) is false);
+
             firstPair.Links.Previous = pair;
             pair.Links.Next = firstPair;
+            pair.Links.Previous = null;
             firstPair = pair;
+
+            Debug.Assert(ReferenceEquals(pair.Links.Next, pair.Links.Previous) is false);
         }
+
+        Debug.Assert(ReferenceEquals(pair, pair.Links.Previous) is false);
+        Debug.Assert(ReferenceEquals(pair, pair.Links.Next) is false);
     }
 
     private KeyValuePair Evict()
@@ -96,6 +110,9 @@ public partial class ConcurrentCache<TKey, TValue>
 
     private void Detach(KeyValuePair pair)
     {
+        Debug.Assert(ReferenceEquals(pair, pair.Links.Previous) is false);
+        Debug.Assert(ReferenceEquals(pair, pair.Links.Next) is false);
+
         if (ReferenceEquals(firstPair, pair))
             firstPair = pair.Links.Next;