+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 Release Notes
 ====
 
+# 05-29-2023
+<a href="https://www.nuget.org/packages/dotnext/4.12.1">DotNext 4.12.1</a>
+* Fixed [162](https://github.com/dotnet/dotNext/issues/162)
+* Fixed other race conditions in `ConcurrentCache`
+* Improved performance of `BitVector` methods
+
 # 03-22-2023
 <a href="https://www.nuget.org/packages/dotnext/4.12.0">DotNext 4.12.0</a>
 * Performance improvements of interpolated string handlers
 
-a href="https://www.nuget.org/packages/dotnext.metaprogramming/4.12.0">DotNext.Metaprogramming 4.12.0</a>
+<a href="https://www.nuget.org/packages/dotnext.metaprogramming/4.12.0">DotNext.Metaprogramming 4.12.0</a>
 * Updated dependencies
 
 <a href="https://www.nuget.org/packages/dotnext.unsafe/4.12.0">DotNext.Unsafe 4.12.0</a>